backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "bcrypt==4.0.1",
    "email-validator==2.1.0",
    "faker>=40.1.2",
    "fastapi==0.109.0",
    "httpx==0.26.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.1.0",
    "pydantic[email]==2.5.3",
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "slowapi==0.1.9",
    "sqlalchemy[asyncio]==2.0.25",
    "uvicorn[standard]==0.27.0",
]
